Refine advertised status endpoint check
