+++ pkg/server/BUILD.bazel
@@ -28,7 +28,6 @@ go_library(
         "//pkg/config/kerneltype",
         "//pkg/domain",
         "//pkg/domain/infosync",
-        "//pkg/domain/serverinfo",
         "//pkg/errno",
         "//pkg/executor",
         "//pkg/executor/mppcoordmanager",
@@ -64,6 +63,7 @@ go_library(
         "//pkg/server/handler/tikvhandler",
         "//pkg/server/handler/ttlhandler",
         "//pkg/server/internal",
+        "//pkg/server/internal/advertisedstatus",
         "//pkg/server/internal/column",
         "//pkg/server/internal/dump",
         "//pkg/server/internal/handshake",
@@ -154,7 +154,6 @@ go_test(
         "conn_stmt_test.go",
         "conn_test.go",
         "driver_tidb_test.go",
-        "http_status_test.go",
         "main_test.go",
         "mock_conn_test.go",
         "server_test.go",
